requirements.txt
pytest
torch
transformers
tokenizers
ir_measures
faiss-cpu
pyserini